moved ComputeControlEnergy and ComputeOptimizedControlEnergy to pipelines.py
